refactor rendering, sizing, and presentation logic
This moves the responsibility of sizing the SVG of the timeline histogram into the ResponsiveSVG component, which wraps the actual SVG in StackChart.js.  The StackChart now owns its own offset controls, which it maintains as state, and all animations are managed by the component MorphPath components.

Still not perfect, but we're getting closer to an actual responsive page